edition module split
