+++ src/app/endpoint/edition/mod.rs
@@ -1,0 +1,12 @@
+mod create;
+pub use self::create::*;
+mod list;
+pub use self::list::*;
+mod delete;
+pub use self::delete::*;
+mod commit;
+pub use self::commit::*;
+
+#[cfg(test)]
+mod tests;
+
