+++ src/app/endpoint/edition/tests/mod.rs
@@ -1,0 +1,7 @@
+mod create;
+pub use self::create::*;
+mod list;
+pub use self::list::*;
+mod delete;
+pub use self::delete::*;
+
